Use regular functions to name commands, not keywords.
The reader no longer just snarfs the first thing on standard input; it.
now peeks at the first character, and tries to read it only if it matches.
a configurable character (*COMMAND-CHAR*). It interns the following symbol.
in the keyword package, then searches the packages in *COMMAND-PACKAGES*.
for same-named symbols; if it finds one, it tries to funcall it.