Commands are now regular functions.
They're regular functions, but we have some special magic for setting.
up the calls to them from the REPL. Our DEFCMD macro now does some.
parsing of the lambda list, pulling out argument reader forms and.
caching them as thunks in the command's plist. At command read time,.
we force the thunks to supply the arguments to the command function.